Fixed a problem in 'compile.py', fixed platform configurations in vc9

+++ src/Win/res/compile.py
@@ -3,6 +3,7 @@
 # Requires gettext (msgcat/msgfmt) and winzip command line addon (wzzip)
 #
 # Revision History
+# 2009-03-05 DRC Fixed some parameter passing issues (spaces/hyphens in names)
 # 2009-03-01 DRC Support multiple po files (by using msgcat)
 # 2009-01-02 DRC Updated to support creation of data files
 """compile.py [-d] executableDir targetname
@@ -16,6 +17,7 @@ import glob
 import os
 import shutil
 import sys
+from subprocess import *
 
 # By default, gettext is in the path, winzip is not.
 # Command line addon to winzip
@@ -33,12 +35,18 @@ def ReadPipe(logFile, cmd):
 		else:
 			break
 
-def RunCommand(command):
+
+# Some commands generate messages on stderr that are interesting.
+# Some are just plain annoying (wzzip)
+def RunCommand(command, toastErr):
 	print 'Running:', command
-	(childin, childout) = os.popen4(command)
-	childin.close()
-	ReadPipe(sys.stdout, childout)
-	childout.close()
+	if toastErr:
+		# Map stderr to a pipe that we ignore
+		p = Popen(command, stderr=PIPE, stdout=PIPE)
+	else:
+		# Map stderr to stdout
+		p = Popen(command, stderr=STDOUT, stdout=PIPE)
+	ReadPipe(sys.stdout, p.stdout)
 
 
 def main():
@@ -69,23 +77,26 @@ def main():
 	for file in ('en', 'fr'):
 		autogen = file + '\\' + autogenFile
 		# -t: output encoding
-		cmd = 'msgcat -t "utf-8" -o ' + autogen + ' ' + file + r'\arb.po'
+		cmd = ['msgcat.exe', '-t', 'utf-8', '-o', autogen, file + r'\arb.po']
 		for po in glob.glob(file + r'\*.po'):
 			if po != file + r'\arb.po' and po != autogen:
-				cmd += ' ' + po
-		RunCommand(cmd)
+				cmd += [po]
+		RunCommand(cmd, 0)
 		installPath = langDir + '\\' + file
 		if not os.access(installPath, os.F_OK):
 			os.mkdir(installPath)
 		# -v: verbose
 		# -c: perform all checks (format,header,domain)
 		# -f: Use fuzzy entres in output
-		cmd = 'msgfmt -v -c -f --strict -o "' + installPath + r'\arb.mo" ' + autogen
-		RunCommand(cmd)
-		if not bDebug:
+		cmd = ['msgfmt.exe', '-v', '-c', '-f', '--strict', '-o', installPath + r'\arb.mo', autogen]
+		# msgfmt generates interesting messages to stderr, don't toast them.
+		RunCommand(cmd, 0)
+		if not bDebug and os.access(autogen, os.F_OK):
 			os.remove(autogen)
 
-	RunCommand('"' + winzip + '" -a ' + executableDir + '\\' + targetname + '.dat DefaultConfig.xml AgilityRecordBook.dtd')
+	# Winzip generates "..." and backspaces while searching for files (stderr)
+	# Toast em!
+	RunCommand([winzip, '-a', executableDir + '\\' + targetname + '.dat', 'DefaultConfig.xml', 'AgilityRecordBook.dtd'], 1)
 
 
 main()
